WIP parser setup

+++ Quokka.RISCV.Integration.Tests/Client/Blinker/Source/externals.h
@@ -1,4 +1,4 @@
 #define l_mem ((volatile unsigned int*)0x00000000)
 #define l_mem_size 512
-#define data ((volatile unsigned short*)0x01000000)
+#define data ((volatile unsigned int*)0x01000000)
 #define data_size 16
